+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Support for primary error message in PHP [#3276](https://github.com/microsoft/kiota/issues/3276)
 - Added support for multiple content type request bodies. [#3377](https://github.com/microsoft/kiota/issues/3377)
 - Added support for multiple content type responses. [#3377](https://github.com/microsoft/kiota/issues/3377)
+- Support for primary error message in Python [#3277](https://github.com/microsoft/kiota/issues/3277)
 
 ### Changed